refactor: add CardViewerFragment
extracts the general functionality of the PreviewerFragment into a more general one
